+++ csrc/huf_wrapper.h
@@ -4,7 +4,7 @@
 #include <stddef.h>
 #include <stdlib.h>
 
-size_t hufCompressData(const u_int8_t *data, size_t size, size_t maxCompressedSize,
+size_t hufCompressData(const u_int8_t *data, size_t size, int maxCompressedSize,
                        u_int8_t *compressedData, size_t *compressedChunksSize,
                        size_t chunkSize, float compThreshold,
                        size_t checkThreshold);
